♻️refactor: 테스트를 하기위한 ParseQueryUtils 분리
