+++ src/main/java/util/HttpRequestUtils.java
@@ -1,16 +1,10 @@
 package util;
 
-import com.google.common.base.Strings;
-import com.google.common.collect.Maps;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.BufferedReader;
 import java.io.IOException;
-import java.util.Arrays;
-import java.util.Map;
-import java.util.Objects;
-import java.util.stream.Collectors;
 
 public class HttpRequestUtils {
     private static final Logger logger = LoggerFactory.getLogger(HttpRequestUtils.class);
@@ -25,35 +19,6 @@ public class HttpRequestUtils {
         }
         logger.debug("request path: {}", path);
         return path;
-    }
-
-    public static Map<String, String> parseQueryString(String queryString) {
-        if (Strings.isNullOrEmpty(queryString)) {
-            return Maps.newHashMap();
-        }
-
-        String[] tokens = queryString.split("&");
-        return Arrays.stream(tokens)
-                .map(HttpRequestUtils::getKeyValue)
-                .filter(Objects::nonNull)
-                .collect(Collectors.toMap(Pair::getKey, Pair::getValue));
-    }
-
-    private static Pair getKeyValue(String keyValue) {
-        if (Strings.isNullOrEmpty(keyValue)) {
-            return null;
-        }
-
-        String[] tokens = keyValue.split("=");
-        // tokens 예외상황 고려
-        if (tokens.length != 2) {
-            return null;
-        }
-        if (tokens[1].contains(".com")) {
-            tokens[1] = tokens[1].replace("%", "@");
-        }
-
-        return new Pair(tokens[0], tokens[1]);
     }
 
     public static String getRequestHeader(BufferedReader br) {
